Add support for turning off connection pooling
This can be necessary in environments that fork new children
after the connection to the database has been established
and stuck in the connection pool. That pool is inherited by
the child but is no longer any good resulting in stale connections.

Apache prefork is such an environment.

By setting postgresql.use_null_pool to True in config, we can force
connection pools to not be used.